+++ src/main/java/plansoft/skillsmanagement/SkillsmanagementApplication.java
@@ -1,13 +1,13 @@
-package plansoft.gestionecompetenze;
+package plansoft.skillsmanagement;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class GestionecompetenzeApplication {
+public class SkillsmanagementApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(GestionecompetenzeApplication.class, args);
+		SpringApplication.run(SkillsmanagementApplication.class, args);
 	}
 
 }
