+++ src/main/java/plansoft/skillsmanagement/area/Area.java
@@ -1,0 +1,37 @@
+package plansoft.skillsmanagement.area;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+
+@Entity
+public class Area {
+	
+	@Id
+	private int id;
+	private String description;
+	
+	public Area() {
+		
+	}
+
+	public Area(int id, String description) {
+		this.id = id;
+		this.description = description;
+	}
+
+	public int getId() {
+		return id;
+	}
+
+	public void setId(int id) {
+		this.id = id;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+}
